[fix][g1full] Fix lost backward references
There is a scenario where a live object happens to be at the start
of the first region in the compaction queue of a thread. There is
a corner case where that object is referenced by another object that
is marked to move to H2, but this object is not marked (eg. it is
metadata).

In that case, inside 'adjust_pointer', the object is in a compacting
region but it does not have a forwardee (it is in the beginning of
the compaction point so it does not move) so the function returns
without marking the card in the H2 Card Table.

If the object moving to H2 ends up in a region without any other
objects with backward references, the card remains clean. This causes
losing the backward reference during the next GC, and the object that
stayed in H1 to be considered dead.